servers/librarian/pyproject.toml
[project]
name = "librarian-server"
version = "0.1.0"
dependencies = [
    "mcp[cli]",
    "pyyaml"
]
